requirements.txt
discord.py==1.3.1
uwuify==1.0.1
toml==0.10.0
click==7.0
git+https://github.com/Rapptz/discord-ext-menus
git+https://github.com/kaylynn234/bigbeans
